+++ R/read_delim.R
@@ -186,7 +186,7 @@ read_delimited <- function(file, tokenizer, col_names = TRUE, col_types = NULL,
     }
   } else {
     if (empty_file(file)) {
-       return(tibble::data_frame())
+       return(tibble::tibble())
     }
     if (is.character(file) && identical(locale$encoding, "UTF-8")) {
       # When locale is not set, file is probablly marked as its correct encoding.