+++ tests/testthat/test-melt-fwf.R
@@ -99,8 +99,8 @@ test_that("melt columns with width, ragged", {
 })
 
 test_that("melt_fwf returns an empty data.frame on an empty file", {
-   empty_df <- tibble::data_frame(row = double(), col = double(),
-                                  data_type = character(), value = character())
+   empty_df <- tibble::tibble(row = double(), col = double(),
+                              data_type = character(), value = character())
    expect_true(all.equal(melt_fwf("empty-file", progress = FALSE), empty_df))
 })