+++ tests/testthat/test-read-csv.R
@@ -203,7 +203,7 @@ test_that("comments are ignored regardless of where they appear", {
   expect_warning(out6 <- read_csv("x1,x2,x3\nA2,B2,C2\nA3,#B2,C2\nA4,A5,A6", comment = "#", progress = FALSE))
   expect_warning(out7 <- read_csv("x1,x2,x3\nA2,B2,C2\nA3,#B2,C2\n#comment\nA4,A5,A6", comment = "#", progress = FALSE))
 
-  chk <- tibble::data_frame(
+  chk <- tibble::tibble(
     x1 = c("A2", "A3", "A4"),
     x2 = c("B2", NA_character_, "A5"),
     x3 = c("C2", NA_character_, "A6"))
@@ -258,7 +258,7 @@ test_that("skip respects newlines", {
 })
 
 test_that("read_csv returns an empty data.frame on an empty file", {
-   expect_true(all.equal(read_csv("empty-file", progress = FALSE), tibble::data_frame()))
+   expect_true(all.equal(read_csv("empty-file", progress = FALSE), tibble::tibble()))
 })
 
 test_that("read_delim errors on length 0 delimiter (557)", {